+++ samples/api.callback.js
@@ -1,8 +1,8 @@
 
-should = require('should');
-stringify = require('../lib');
+const stringify = require('../lib')
+const assert = require('assert')
 
-input = [ [ '1', '2', '3', '4' ], [ 'a', 'b', 'c', 'd' ] ];
+input = [ [ '1', '2', '3', '4' ], [ 'a', 'b', 'c', 'd' ] ]
 stringify(input, function(err, output){
-  output.should.eql('1,2,3,4\na,b,c,d\n');
-});
+  assert.equal(output, '1,2,3,4\na,b,c,d\n')
+})