+++ samples/api.pipe.js
@@ -1,8 +1,11 @@
 
-var stringify = require('../lib');
-var generate = require('csv-generate');
+const stringify = require('../lib')
+const generate = require('csv-generate')
 
-generator = generate({objectMode: true, seed: 1, headers: 2});
-stringifier = stringify();
-
-generator.pipe(stringifier).pipe(process.stdout);
+generator = generate({
+  objectMode: true,
+  seed: 1,
+  headers: 2
+})
+stringifier = stringify()
+generator.pipe(stringifier).pipe(process.stdout)